Add message limit for sending emails
Telegram has a limit of 4KB per messages.
we could split the message and send it in multiple
parts, but I think that is not an important case as
users notice the message so if they want they can
check it in their emails.
The reason we limited that to the length of 2000
is because of the UTF-8 that each character in persian
takes 2 bytes. We could check that with reflect of unsafe
package, but as i mentioned that is not an important case
to work on it